A solution in J.
Ok, kind of late to the party.  I stumbled across a description of J,
and thought that it sounded interesting, and that it actually probably
would be pretty good at this sort of thing, given its APL heritage.  I
won't claim that this is idomatic J.  But actually, after writing it
and pruning it a bit... no, it still is pretty cryptic.